separation of js
